password generator
In this script, we define a function generatePassword that takes a single parameter length which specifies the length of the password to be generated. The function creates an empty string result, and a string characters which contains all the possible characters that can be used in the password. We then get the length of the characters string using the length property.

We then use a for loop to iterate length number of times, and each time we append a random character from the characters string to the result string using the charAt method and the Math.random() and Math.floor() functions.

Finally, we return the generated password.

In the last line, we call the function and pass the value 8 as the length parameter. This generates and logs a random password of length 8 to the console.